[apic_aim] Map neutron resources to AIM, part 3
Implements an L3 service plugin that, in conjunction with the apic_aim
mechanism driver, maps each Neutron router to an AIM Contract and
ContractSubject whose DNs and status are exposed via extended
attributes similar to those on the core Neutron resources. An "any"
Filter and FilterEntry are created per-tenant, and referenced in this
contract, allowing all traffic from EPGs providing and consuming this
contract to be routed.

The add_router_interface and remove_router_interface methods are stubs
that will be implemented in the next patch set. They will manage the
mapping of router interfaces to AIM Subnets, along with having the
default EPGs associated with those interfaces provide and consume the
router's Contract.

The corresponding GBP policy driver's extension is renamed
apic_aim_gbp for consistency with the apic_aim and apic_aim_l3
extensions at the Neutron level, and all extensions are now in the
gbpservice.neutron.extensions module.

The GBP policy driver's unit tests are updated to account for the
Filter and FilterEntry resources created by the mechanism driver.

Change-Id: I1b7f0c80e66d55d58c27fe9e4cb461f62aec3c42