[apic-mapping] Automatic PTG per L2P
This change automatically creates a PTG (with the name "auto-ptg-<l2p_id>")
per L2P. This PTG is created as a reverse map of the "shadow" EPG that was
already being created per L2P.

In order to maintain the reverse-mapping integrity between the EPG and the PTG,
an entry is created in the name-mapping DB that maps the ID of the auto PTG
to the "apic-name" of the "shadow" EPG.

The ID of the auto PTG is derived from the ID of the L2P as a MD5 hash
calculation (for uniqueness) and persisted in the format "auto<hash>". It is
thus always possible to determine the ID of the auto PTG from the ID of the
L2P and no additional state needs to be maintained.

The auto PTG cannot be deleted by the end user and doing so will result in
an error.

The user can update the name, description, provided and consumed PRS for the
auto PTG, but cannot update any other attributes and doing so will result in
an error.

The shared status of the auto PTG is made consistent with the shared status
of the L2P (once set, it cannot be changed).

The auto PTG is deleted when the corresponding L2P is deleted (attempted in
the pre-commit phase).

To prevent forward mapping of the auto PTG to a new EPG, all above
operations are invoked on the GBP DB mixin (parent of the GBP plugin). This
ensures that the apic_mapping policy driver is not invoked for the create and
delete auto PTG operations during L2P creation and deletion.

The creation of the auto PTG is controlled by a configuration and is disabled
by default thus allowing this new feature to be turned ON only where needed.
All existing deployments should not see any change in behavior as long
as they choose not to turn ON this feature. This configuration is as follows:

[apic_mapping]
create_auto_ptg=<True or False>

Change-Id: Ie132ace0fc9f78baa0034a6f30f2ee758bb271c0