[apic-mapping] Default PTG per L2P
This change creates a default PTG (with the name "default-ptg") per L2P. This
PTG is created as a reverse map of the "shadow" EPG that was already being
created per L2P.

In order to maintain the reverse-mapping integrity between the EPG and the PTG,
an entry is created in the name-mapping DB that maps the ID of the default PTG
to the "apic-name" of the "shadow" EPG.

The default PTG cannot be created or deleted by the end user and doing so will
result in an error.

The user can update the provided and consumed PRS for the default PTG, but
cannot update any other attributes  and doing so will result in an error.

The shared status of the default PTG is made consistent with the shared status
of the L2P (once set, it cannot be changed).

The default PTG is deleted when the corresponding L2P is deleted (attempted in
the pre-commit phase).

To prevent forward mapping of the default PTG to a new EPG, all above
operations are invoked on the GBP DB mixin (parent of the GBP plugin). This
ensures that the apic_mapping policy driver is not invoked for the above
operations.

The creation of the default PTG is controlled by configuration and is disabled
by default thus allowing this new feature to be turned on only when and where
needed. All existing deployments should not see any change in behavior as long
as they choose not to turn on this feature. The configuration is as follows:

[apic_mapping]
create_default_ptg=<True or False>

Change-Id: Ie132ace0fc9f78baa0034a6f30f2ee758bb271c0