Removed DVS agent RPCs
This removes the RPCs to the DVS agent. The RPCs were needed
in the APIC ML2/Neutron workflow for port binding in order to
allow security groups to be implemented by the DVS agent. However,
security groups aren't used in the GBP workflow, and therefore these
RPCs aren't needed and should be removed.

closes-bug: 1621617

Change-Id: I16a5ba64828a71ac51bdc1b4669dfba62fb45f07
Signed-off-by: Thomas Bachman <tbachman@yahoo.com>