[apic_aim] Map neutron resources to AIM, part 4
The apic_aim mechanism driver and L3 plugin map router interfaces to
AIM Subnets. The DNs and status of these subnets are exposed via
extended Neutron subnet attributes. If any subnets of a network are
attached as interfaces to a router, the network's default EPG provides
and consumes the router's Contract.

A seperate patch will manage of the VRFs of routed networks, and will
reject invalid routing topologies, completing the basic east/west
routing functionality.

Another follow-on patch will additionally expose the AIM Subnets via
extended attributes of the router to with the corresponding Neutron
subnet is attached, and will likely include both the Neutron subnet
and router names in the AIM display name of the Subnet.

Change-Id: Id8aa749c2a590bf6d0548162483553edb8a3589d