[apic-mapping] Segmentation label extension for PT
This patch defines a new extension: apic_mapping_extension, for the
apic_mapping policy driver. An extension attribute:
segmentation_labels, that extends the Policy Target definition, is
being introduced in this extension.

A corresponding extension driver: apic_mapping_extension, that processes
this extension, is also being added.

The segementation_labels attribute is a list of strings. Each string can
be upto 1024 characters long. These labels are not interpreted by GBP
but are instead passed downstream by the apic_mapping driver. It is
assumed that these are defined outside of OpenStack and the backend
system can appropriately interpret them.

The get_gbp_details() RPC call implemented by the apic_mapping driver
will return the segmentation_labels in its body.

A CLI option: --segmentation-labels will be provided for the
policy_target create and update operations. This CLI option will accept
a comma separated string as the option value.

Change-Id: I360bf9f7f1d4bdca76d4f16b7535a6416f430830