[WIP][apic_aim] Map neutron resources to AIM, part 5
Complete basic east/west routing. Enables routing on BDs of routed
networks and associates those BDs with the address scope's VRF if
applicable, or else with a per-tenant default routed VRF.

This is currently a work in progress. The chosen VRF still needs to be
exposed via the extended attributes of the Neutron network and router
resources. Validation of routing topologies is still required as well.

Change-Id: Ic7396e5ebbc466ea5be0028931b31bdbab9833e6