AIM Policy Driver - Part 6 - Add Router Interfaces
This patch adds interface for the PTG subnets to the default Router.
Corresponding UTs are added which check that the correct port is created for
the router interface.

A bug due to which the apic_aim_l3 was not getting set for the GBP aim_mapping
tests is also being fixed.

Change-Id: Ia8e96da15b2571491412c649af5a99261ceb8a84