[aim-mapping] Fix PT create error in auto-ptg
The commit 6d56931196b6c78663a506892761fd9fe86373d5 introduced a
variation in the UUID of the "auto" PTGs implicitly created by the
apic_mapping policy driver. The API layer however was not made aware
of this variation and hence rejected this new UUID format when used
to create a policy target. This is fixed in this patch by allowing
the validation for this new UUID format with the regex:
auto[0-9a-f]{32}\Z

The auto PTG was also missing an implicit subnet for IP address allocation
to any PTs that are created in that PTG. This is also being fixed.

UTs have been added/updated to validate the above changes.

Change-Id: Idda35feb5c61587f3f014491768daecf03660ad9
Closes-bug: 1630923