[aim-mapping] Queue DHCP and Nova notifications
Currently, whenever resources are created by GBP (either GBP resources or
Neutron resources), DHCP and Nova notifications are sent immediately
after a specific resource creation completes, i.e. the transaction
handling that resource is complete. However, this transaction might
actually be a sub-transaction that is nested within other transactions
and the resource state is committed to the DB, only when the outermost
transaction completes.

This creates a problem in the aim-mapping policy driver which wraps every
user-facing operation in a big nested transaction, but the notifications
for implicitly created resources are sent before the outermost transaction
is committed.

The bug in question is one such manifestation of the problem where the
subnet create notification is sent to the DHCP agent as a part of the PTG
creation orchestration. The DHCP agent tries to validate the subnet by
checking the presence of the associated Neutron network, but neither the
subnet nor the network is yet committed to the DB since the outermost
transaction for the PTG creation is not yet complete.

This patch introduces a queueing framework for queueing DHCP agent and Nova
notifications, and then sending them once the outermost transaction (parent)
completes (by registering for Sqlalchemy transaction event hooks). The order
of the notifications is maintained when the notifications are eventually
dispatched.

This queueing scheme is turned ON only for the aim-mapping driver (as a part
of its initialization). Any other policy drivers wanting to leverage this
queueing feature should set the following attribute:

BATCH_NOTIFICATIONS

of the:

gbpservice/network/neutronv2/local_api.py

module to True.

Change-Id: I23a12eda7f98970431c8cbea4acad18f762b201d
Closes-bug: 1631833
(cherry picked from commit e553a03e23ac65b1a5a57e0a75d0233528bf9538)