[apic-aim] Fix unit tests for AIM update
Now that AIM keeps resources in PENDING state until AID actually
synchronizes them, as it should, the TestSyncState.test_*_synced tests
need to mock get_status to return the SYNCED state.

Also, a new attribute added to the AIM resources was breaking UTs for
the aim-mapping driver.  This new attribute does not need to be
validated and is fixed here.

Change-Id: I84f3181ba769c58773137b34f858888d939f3fcc