Propagate relevant driver exceptions to client
We have so far taken the approach of morhping the driver specific
exceptions into generic GBP exceptions before we propagate them
back to the client. The intention here was to shield the end user
from the technology specific driver details.

Over time, we have seen that the generic GBP exceptions make the
system difficult to debug, often requiring painful correlation
between client request-id and the server side logs.

This patches proposes to err on the side of exposing more details
in the form of the driver exception but hopefully making it easier
to debug an error situation.

Closes-bug: #1636624

Change-Id: Ia067b4bfa9d155ab3d780e1e69c8e5ceedd75cd6