[AIM] Add support for SNAT
To support 'distributed' SNAT, this change
adds a way to allocate an IP-address per host
that will be used for translating traffic leaving
VMs on that host. This IP-address is conveyed
to the host through GBP RPCs.
Specific subnets in the external network can be
designated for use as a pool of SNAT addresses
using the boolean extension attribute
'host_snat_pool'. Such subnets cannot be used
for allocating floating-IPs to maintain
consistency with current APIC driver, but that
can be relaxed if required.

Change-Id: Ie04095be77a6aefcf8559b743d265fe66551c2bd
Signed-off-by: Amit Bose <amitbose@gmail.com>