Reduce log level for UTs
At least one deprecation message at WARN log level in Neutron is leading
to excess logging in the UTs, which in turn leads to log files > 50 MB and
causes the py27 gate job to fail. To workaround this (only in this
stable/liberty branch), the log level is being set to ERROR in the GBP base
test.

In addition the an environment variable is set to Ignore deprecation warnings
in UT runs.

Change-Id: I70ca386214c12a65539ae8d2935e3c4fb41c08d0