Change expunge_all to expire_all
Change ID Ic9ccdfbf06cf1c5e1ca1cc648eedce2d09c76540 introduced
the use of expunge_all to ensure consistent state when a session
is used for consecutive transactions. The expire_all can be used
instead, which allows objects in the persistent state (i.e. still
tied to a session) to still be used, and ensure that they refetch
the state from the DB.

Closes-Bug: 1642372

Change-Id: If5cc4f842714136f54ff5530bc43db34f8121f8c
Signed-off-by: Thomas Bachman <tbachman@yahoo.com>