[AIM] Eliminate name mapping DB
AIM resource names are now Neutron and GBP resource UUIDs, with an
optional prefix.

This is the first step of phasing out the name mapper. Once an
appropriate function, similar to aim.utils.sanitize_display_name, is
available from AIM, calls to the name mapper should be replaced with
calls to that function. Once this is complete, the name mapper module
and class will be removed.

Change-Id: I6b4b180f8722e4378e3afee7d62a71292eea233d