Change passing session to context in segments db functions
We have a number of functions that expect to get session as one
the arguments. Passing session is not correct and this prevents
of usage new enginefacade which expect context to be passed and
session will be injected in current context.

This change is based on change in Neutron, so here all affected
places are fixed to pass context.

Change-Id: If55d11b6156a3cfec83f9ea4547a38019c4deb22
Depends-On: I68ac442f11a6698bc43210fae877ba3809b1c02d