NFP - Support module priority
Enhanced NFP framework to support priority based
rpc loadbalancing. NFP modules could have multiple
rpc managers for a single topic with but with
different priorities.
Oslo does not understand priority,
hence nfp is enhanced to sort managers in priority
order and register with oslo, so that when oslo
logic searches for a rpc handler, it always gets
the higher priority handler.

Change-Id: I148add0837505ed38a72942a1e0f33a1374853ed
Closes-Bug: 1644283