[apic_aim] Address scope sharing
The apic_aim mechanism driver now implements sharing of
address_scopes, along with sharing of networks associated with
address_scopes. Sharing of unscoped networks will be implemented in a
subsequent patch. The aim_mapping GBP policy driver will also need to
be updated in a subsequent patch to support shared address_scopes.

When a Neutron network is interfaced to a router and associated, via a
subnetpool, with an address_scope, the AIM BridgeDomains,
EndpointGroups, and Subnets for that network are now mapped under the
AIM Tenant of the address_scope. Router Contracts are now mapped under
the common AIM Tenant, along with the AnyFilter used by the router
Contracts.

Attempting to add an IPv6 subnet to a router now results in an
IPv6RoutingNotSupported exception, and attempting to interface
networks associated with different address_scopes to the same router
now results in a MultiScopeRoutingNotSupported exception. IPv6 routing
and multi-scope routing will be implemented in subsequent patches.

Attempting to change the address_scope_id of a subnetpool from which
subnets are interfaced to routers currently results in a
ScopeUpdateNotSupported exception. This exception is currently
replaced by ML2Plus with a MechanismDriverError exception, which may
be corrected in a subsequent patch. Proper handling of updates to a
subnetpool's address_scope_id attribute will be implemented in a
subsequent patch.

Change-Id: I04fc9ae47de6aebe8cd7793eed958abfa8b5104f