[aim-mapping] Fix PTG lifecycle regression
Changes introduced in the patch to create the auto-PTG caused
some regressions with the PTG workflows -

1. The call to create_policy_target_group() plugin method (instead of
the DB plugin) caused an earlier loaded PTG DB object to go out of scope.
This is being fixed by repositioning where the PTG DB object is loaded.

2. Since auto-PTG is not maintained as an implicitly created resource,
the L2P was not getting deleted. A new check is added to allow
deletion of the L2P if the auto-PTG is the only PTG present (deletion
of the L2P can still fail if the auto-PTG cannot be deleted). While doing
this an issue with loading the L2P DB object was encountered and is being
fixed by an explicit session.refresh call.

Change-Id: Ic9e586d4ef31b927770dca0eb755cc85629edb01
Closes-bug: #1647878