[aim-mapping] intra-ptg-allow ext attr for PTG
This attribute maps to the policy_enforcement_pref attribute
of the AIM EPG, and is used to turn ON or OFF the intra-PTG
traffic. It's True by default, implying all PTs in a PTG are
allowed communicate with each other. When set to False, there
is no direct connectivity between the PTs of a PTG.

Change-Id: I82094cbc1faaed879bfb75518df0f558f0bdad1c
Closes-bug: #1648019