[aim] Fix aim gate job
This patch fixes the gate-group-based-policy-dsvm-aim-ubuntu-xenial-nv
integration job that exercises the aim_mapping policy driver and
the ml2plus plugin with the apic_mapping mechanism driver.

Other GBP integration jobs patch a devstack installation to run GBP, however
this patch configures the aforementioned job to use the in-tree GBP devstack
plugin. Based on the validation from this patch, the same approach can now be
extended to the other jobs as well and has the advantage of providing the
developer the ability to recreate the devstack setup used in the gate job.

Some variations do exist in the devstack setup on account of the environment
available in the gate job versus that in a developer's setup (for example, the
source repo is already cloned in the gate job and needs to be used), and in
such cases the ENABLE_APIC_AIM_GATE devstack env variable is used to
special case appropriately. Anyone wanting to replicate the exact devstack
gate configuration should set this variable to true in their local.conf
instead of the ENABLE_APIC_AIM variable referred to in the devref document
and which adds more configuration and starts neutron agents. Over time, the
ENABLE_APIC_AIM_GATE configuration will converge with that of the
ENABLE_APIC_AIM and the ENABLE_APIC_AIM_GATE variable will be eliminated.

The eventual goal of this job is to be able to test as much of the GBP/Neutron
workflowsi, and resulting AIM and system configuration, as possible (in spite
of the absence of the backend components). The first step towards that end
is to test the CRUD operations. This is achieved by running a basic GBP exercise
script that is a variant of the one run in the
gate-group-based-policy-dsvm-functional-ubuntu-xenial-nv job (it does not launch
the VMs since it will fail on account of the expected port binding failure).

The gbpfunctest functional test suite that is run against the
gate-group-based-policy-dsvm-functional-ubuntu-xenial-nv is also run in this
job with some minor adjustments to the test suite. The adjustments are mainly
to account for the variation in sharing semantics between the aim_mapping and
resource_mapping drivers (the former has more flexibility that the latter), and
the lack of support for hierarchical PRS in the aim_mapping driver.

All neutron agents are being disabled in this iteration to avoid logs being
inundated with port binding failure messages. A number of options are up for
consideration in terms of how to get the port binding to succeed, and will
be evaluated over time. The appropriate agents will be enabled accordingly.

Change-Id: I528b7fd2e66a0f7ef1bb00c887e5fbf2434cf966