[wip][aim-mapping] Restrict auto-ptg access
This patch restricts GET and UPDATE for auto-ptg to the admin
via policy.json mechanism. When these rules are in effect, policy_target
creation is also restricted to the admin. These rules can however be
relaxed if required by modifying the policy.json file (information
on how to do it is documented in the policy.json that is packaged with
this project).

This also adds a new is_auto_ptg attribute to facilitate specification
of rules in policy.json. This has the added benefit of providing a way
to filter the policy_target resource to retrieve auto_ptgs.

Change-Id: I6d9e873acb2b1b3bee8d78a45527bd4d5d437eca