[apic_aim] Unscoped network sharing
The apic_aim mechanism driver now implements routing of shared
unscoped networks. If any network in an unscoped topology is shared,
the AIM BridgeDomains, EndpointGroups, and Subnets for all networks in
that topology are now mapped under the same AIM Tenant as that shared
network's BD and EPG, and that AIM Tenant's default routed VRF is
used.

Attempting to join topologies that include networks shared by projects
mapping to different AIM Tenants results in an exception.

Management of external connectivity for shared unscoped networks will
likely require updating in a followup patch, and notifications may be
required when topologies are moved between VRFs. Also, an additional
patch will be needed to ensure that overlapping subnets are not
allowed within a VRF when multiple independent unscoped router
topologies use the same AIM Tenant's default VRF.

Change-Id: I220354aa15448aa85aa17712ed4715b3d9017021