[WIP][aim-mapping] Resource sharing for GBP resources
The Contract, ContractSubject, Filter, and FilterEntry AIM resources
mapping to the GBP PolicyRuleSet and PolicyRule are now being always
created in the "common" ACI tenant to implement shared semantics for
the corresponding GBP resources. The value of the shared flag for
these GBP resources will control the visibility of the GBP resources,
and thus the usage for the mapped AIM/ACI resources, although by
virtue of being created in the common tenant, the ACI resources will
always be available for sharing.

The testing strategy in the GBP UTs is now changed to filter AIM
Filter and FilterEntries relevant to a tenant (by examining the
tenant_id encoded in their names) and only validate those. The
Filter and FilterEntries created by the apic_mapping mechanism
driver are thus flitered out and no longer need to be accounted
for (thus insulating these UTs from further changes in the
apic_mapping MD).

The display_name for the implicitly created Filters is being set
in a relevant way in cases where it was earlier showing up as a
static string "per-project". In such cases its now being to set
to the name of the resource.

The L3P's shared status needs to match the shared status of the
address_scope and subnetpool(s) it's associated with. When the
L3P is created with explicit address_scope, and/or subnetpool(s)
the L3P's shared status needs to be derived from the address_scope.
This was correctly happening in the case of explicitly provided
address_scope, but not in the case of explicitly provided
subnetpool(s). This is being fixed here. UTs have been updated
to run all create L3P scenarios with both shared and unshared
options. Since the L3P does not directly map to an AIM resource,
it is expected that the apic_mapping MD will appropriately handle
the address_scope and subnetpool(s) shared attribute and the mapping
to AIM resources.

Updating the shared attribute of the L3P is not supported. This is
consistent with the behavior of the earlier apic_mapping policy
driver.

In the implicit resource mapping workflow, the L2P's shared status
needs to be set per the shared status of the network its associated
with. This was already being done, but is now being validated in UTs in
this patch.

Since the L2P does not directly map to an AIM resource,
it is expected that the apic_mapping MD will appropriately handle
the network shared attribute and the mapping to AIM resources.

When the L2P is created with the explicit network, a non-shared
explicit network cannot be associated with a shared L2P. This check
is added here. It should be noted that if the L2P is created
without a refernce to a L3P, the L3P will be created implicitly (or
an earlier created default L3P will be used) and will be associated
with an address_scope. Any subnets which are associated with the
explicit network should be associated with the subnetpool and
address_scope associated with the L3P. This means that if there are
any prior subnets associated with explicit network, they will not fail
the subnetpool and address_scope association validation check in the
Neutron layer and the L2P creation will fail.

The _aim_tenant_name method has been updated to compute the appropriate
tenant for a BridgeDomain. The appropriate tenant is the tenant of the
L3P.

In the case of a PTG, the mapped EndpointGroup needs to be in the
appropriate tenant regardless of the shared status of the PTG. The
_aim_tenant_name method has been updated to compute the appropriate
tenant for a BridgeDomain. The appropriate tenant is the tenant of the
L3P for this PTG.

Change-Id: Ic1c5d24a0b8aedb3b8f1a24a76c9c98854163321