Fix NSP UT failure in Resource Mappind Driver
The following Neutron commit changed the lazy loading scheme
for Networks and Subnets:
https://github.com/openstack/neutron/commit/459437c5a6bcbd54cdb0de26e47eb199f60fc734

This is causing the allocation_pools for subnets to not load correctly.
The issue will likely not happen if transactions are used appropriately,
however that may not be happening at least in some places in the RMD. This
patch provides a workaround by expiring the session objects so that
the retrieval subnets is fixed (the cause of the UT failure). The breakage
and fix is specific to the RMD (the issue has not been see elsewhere).

Closes-bug: #1654109

Change-Id: If80cae54d6c8a56e2c01fc88ee8a4720c99c2709