Skip unrouteable subnets with Edge NAT
With GBP workflow, the unrouteable subnet for the external
segment should not be used as the primary IP on the ASR's external
VLAN sub-interface (i.e. the one that connects to the outside world).

This needs to be fixed, such that a NAT pool must be configured/added
to an external segment before an L3P is associated with an L3 segment
(Edge NAT mode only). An IP address from this NAT pool will then be used
for the primary IP on the external VLAN sub-interface of the ASR.

This patch enforces the requirement for a NAT pool to be configured
when using Edge NAT.

Change-Id: I1ce634747efab35b3cf559796594b6782c77c3e9
Closes-Bug: 1654119