[AIM] Allow creating address-scope with pre-existing VRF
This change allows address-scope creation to
accept an additional attribute whose value is the
DN of a VRF that the address-scope should map to.

As a special case, if the pre-existing VRF is in
the common tenant, then moving BDs/EPGs between
tenants due to sharing is not needed.

Change-Id: I69cc3607268f8f256b0a54c90ccf8f7950ef58a9
Signed-off-by: Amit Bose <amitbose@gmail.com>
(cherry picked from commit 1985f82211458c565ffb2fca781f5e0100808ed1)