[aim-mapping] NAT-pool & NSP support
This patch updates the aim_mapping driver to support
the nat_pool and network_service_policy resources. Since
the implementation of these resources is in the context of
Neutron objects (namely, external networks, subnets, and
floating_ips), the realization of these in the aim_mapping
policy driver is similar to that in the resource_mapping
driver with the exception that all post_commit operations in
the resource_mapping driver now need to be accomodated in the
pre_commit operation of the aim_mapping driver. This patch
accordingly adapts the resource_mapping driver's implemantation
to the aim_mapping driver (mostly by refactoring existing code).

New UTs have been added for the aim_mapping driver to test these
two resources, but these UTs have the same strategy and coverage as
that for the resource_mapping driver (they are essentially
refactored copies of the UTs for the resource_mapping driver).

Implementing the above UTs revealed some bugs which are being
also being fixed, and which were required to get the UTs to
work. When two floating_ips were being deleted in the UT:
test_update_nsp_nat_pool_after_pt_create
only one was being shown deleted when a list was done. This
was on account of the nova notifier creating a new session
after deletion of a floating_ip. This led to the realization
that the batching of the update and delete notifications was
not being done (as was earlier introduced for the create
operation).

Change-Id: I6727aacfa5aae0c47034f68dc915257093815cdf