[aim-mapping] NAT-pool & NSP support
This patch updates the aim_mapping driver to support
the nat_pool and network_service_policy resources. Since
the mapping of these resources is to Neutron objects, the
realization of these in the aim_mapping policy driver is
similar to that of the resource_mapping driver with the
exception that all post_commit operations in the
resource_mapping driver now need to be accomodated in the
pre_commit operation of the aim_mapping driver. This patch
accordingly adapts the resource_mapping driver's implemantation
to the aim_mapping driver (most by refactoring existing code).

New UTs have been added for the aim_mapping driver but these
UTs have the same strategy and coverage as those for the
resource_mapping driver (they are essentially refactored copies).

Change-Id: I6727aacfa5aae0c47034f68dc915257093815cdf