[apic-mapping] Call the mapper properly to get the right PTG name
This is to make sure we can get the PTG name stored in the mapper DB
while doing the bind_dvs_port(). Otherwise when the PTG name is
chagned on openStack side, we will use the new PTG name which is
wrong as APIC and vCenter are still using the old name.

Partial-bug: 1547723
Change-Id: Iec7f614e5d84c1d51d7ad7eff6bf6d89eb362fbc