[AIM] Send port notifications on router changes
When a router that has external connectivity is
connected/disconnected from a subnet, the ports
on that subnet may gain/lose external connecitivity
via SNAT. To reflect this on the data path, port
update notifications need to be sent out so
that the opflex agent may refetch information
about the port.
The same applies for external-gateway changes
of a router, except that the set of ports affected
is all the ports in all the subnets that are
connected to the router.

All these notifications are not sent immediately
but delayed until the current database transaction
completes.

Change-Id: I346159ed4217f6ff3442d5c3a75e54ea7841ee29
Signed-off-by: Amit Bose <amitbose@gmail.com>