Making notification-queue use thread-safe
Create notification queue per session and associate with
the session object. This helps us to get away from having
to create a global notification queue.

This at least helps to address the issue seen in commit:
https://github.com/openstack/group-based-policy/commit/bcadb5a54f0680f4c94462dbb8d89c811cdc61f5

which caused certain tests to fail. An additional check was
added in that commit to see if the LOCAL_API_NOTIFICATION_QUEUE
variable was uninitalized but this should have never been the case.
This led to the suspicion of unsafe thread behavior.

Change-Id: Iddd9c7b9f8975d37ccd00d28776b982b90f51324