Introduce implicit subnetpool extension
In order to preserve legacy scripts that create Neutron resources such
as Networks or Subnets, we need a way to specify default address pools
that can be assigned to subnets upon creation.

Something similar already exists in Neutron: an attribute "is_default"
that can be set on a subnetpool so that it can be used by subnets that
don't specify a subnetpool themselves. This however has a bunch of
limitations such as:

- For reasons mentioned by Armando in [0], the intent of using the
  default subnetpool has to be explicit at subnet creation;
- When a subnetpool doesn't exist for a Subnet requesting the default
  one, an exception is raised instead of leaving the subnet without
  any subnetpool;
- Only one default subnetpool can exist in the whole system regardless
  of its visibility (shared, non shared).

To accomodate our use cases, we are introducing a new extension for
the subnetpool resource called "last_resort_subnetpool" and a new
attribute for the subnetpool API "is_last_resort". When a subnetpool is
marked as implicit the following semantics apply:

- Only one implicit subnetpool can exist *per tenant* (if not shared)
- Only one shared implicit subnetpool can exist in the whole system.
  This doesn't affect private implicit subnetpools
- Whenever a subnet is created, the current subnetpool retrival
  mechanism is executed. If that returns no result, a implicit
  subnetpool will be searched in the Subnet tenant. If that fails too,
  we fall back to a global implicit subnetpool. In case all of this
  fails, the subnet will be created with no subnetpool as a legacy
  resource.
Change-Id: Iead6ad458cdcd8ca7c79f5df553336d6147e4c62